+++ main.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 
 	"github.com/judimator/augurken/cmd"
-	"github.com/judimator/augurken/formatter"
+	"github.com/judimator/augurken/log"
 )
 
 var exitFn = os.Exit
@@ -13,11 +13,10 @@ var exitFn = os.Exit
 func main() { exitFn(run()) }
 
 func run() int {
-	logger := formatter.NewLogger()
-	command := cmd.NewCommand(filepath.Base(os.Args[0]), logger)
+	command := cmd.NewCommand(filepath.Base(os.Args[0]))
 
 	if err := command.Execute(); err != nil {
-		logger.Error(err)
+		log.Error(err)
 
 		return 1
 	}