+++ CHANGELOG.md
@@ -1,10 +1,28 @@
+## [v1.3.0](https://github.com/judimator/augurken/tree/v1.3.0)
+
+**Enhancements:**
+- Handle command errors in another way
+- Add golangci-lint into GitHub workflow
+
+## [v1.2.0](https://github.com/judimator/augurken/tree/v1.2.0)
+
+**Enhancements:**
+- Add check command
+- Add tests
+- Refactoring
+
+## [v1.1.3](https://github.com/judimator/augurken/tree/v1.1.3)
+
+**Enhancements:**
+- Add tests
+
 ## [v1.1.2](https://github.com/judimator/augurken/tree/v1.1.2)
 
 **Enhancements:**
 - Add tests to Github workflow
 - Add CHANGELOG.md
 
-## [v1.0.0](https://github.com/judimator/augurken/tree/v1.1.1)
+## [v1.1.1](https://github.com/judimator/augurken/tree/v1.1.1)
 
 **Enhancements:**
 - Add Dependabot
@@ -14,8 +32,8 @@
 **Enhancements:**
 - Compact JSON in "Examples" table
 
-## [v1.1.1](https://github.com/judimator/augurken/tree/v1.0.0)
+## [v1.0.0](https://github.com/judimator/augurken/tree/v1.0.0)
 
 **Enhancements:**
 - Initial commit with the formatter
-- Add Github workflow
+- Add GitHub workflow